Add files via upload
BACKGROUND 
Mash, Who is a pizza lover, establish a pizza restaurant with the aims of providing pizza to his city and his fellow pizza lovers. After a year of embarking into the business, he gathered sufficient data and knew the importance of data in today’s business world, but was clueless about how it could drive his business. He reached out to me,  a data analyst, to help him provide answers to the business questions that linger his mind.
The business questions that were provided are as follow;
1)	What is the Total revenue?
2)	What is the Average order value?
3)	What is the Total pizza sold?
4)	What is the Total order placed?
5)	What is the Average pizza per order?
6)	What is the Daily trend for total orders?
7)	What is the Monthly trend for the total orders?
8)	What is the Percentage of sales by pizza category?
9)	What is the Percentage of sales by pizza size?
10)	What is the Total pizza sold by pizza category?
11)	What is the Top 5 best sellers by revenue, total quantity and total orders?
12)	What is the Bottom 5 best sellers by revenue, total quantity and total orders?

ABOUT THE DATA SET
Mash provided the data set in a table for the analysis containing; Pizza Id, Pizza name, Pizza size, Pizza ingredients, pizza category, date of order and units price for each pizza